Solved several FindBugs issues

+++ app/src/main/java/at/conradi/helloworld/DiceSpinnerActivity.java
@@ -38,8 +38,8 @@ public class DiceSpinnerActivity extends AppCompatActivity {
         @Override
         public void onGlobalLayout() {
             this.layout.getViewTreeObserver().removeOnGlobalLayoutListener(this);
-            float width = layout.getMeasuredWidth()/2;
-            float height = layout.getMeasuredHeight()/2;
+            float width = layout.getMeasuredWidth()/(float)2;
+            float height = layout.getMeasuredHeight()/(float)2;
 
             createSimpleSpinningDice(this.layout, width, height);
         }
